// src/app/models/contact/contact.model.ts
export class Contact {
    id?: any;
    firstname?: string;
    lastname?: string;
    phone?: number;
    email?: string;
    address1?: string;
    address2?: string;
    city?: string;
    state?: string;
    country?: string;
    zipcode?: number;
    active?: boolean;
    grouptitle?: string;
}